Change Interface to be Disjoint from Record
Following suggestions from @darthscsi and @mwachs5, change the actual
"interface" to not be a Record, but to be a separate type.  Demonstrate
how information that is only knowable by an implementation, e.g., a
clock frequency, can be broken out into a type class-ish pattern that
relies on knowing an Interface conformance.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>