Add sketch of how parameters work
Add a function to the Interface that can store parameters, i.e., values
that are passed into a conforming implementation.  This currently is
marked "final" because overriding it will cause link errors due to the
fact that Chisel does not support modules with parameters.  This will be
revisited in the future.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>