Rework Parametric Example to Inject Interface
Rework the parametric example to make the Interface the big boss.  There
is a problem where the parameters of the Interface are chosen by the
client.  This doesn't really work.  Instead change the client to use a
dependency injection pattern to have it take an Interface as a
parameter.  This enables a concrete, Scala-monomorphic Interface object
to be used to drive the creation of the client.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>