Simplify all FIRRTL compiles to use compile-n dirs
Change FIRRTL to Verilog compilation to output all results in
directories of the form "compile-$n".  Previously, the first compilation
would be output in the top-level build directory and subsequent ones
would be placed in "compile-$n".  This simplifies things for situations
where the first thing compiled may not be the top module.  (This can
concretely happen when using Definition/Instance.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>