Add Definition/Instance Example
Add example derived from Adam's Scastie showing Definition/Instance used
for separable compilation.  This shows using the results of the
elaboration of a component in a dependent, second elaboration of client
that instantiates the component twice.  This was designed to have the
same structure as the Interface series of examples.

Co-authored-by: Adam Izraelevitz <adam.izraelevitz@sifive.com>
Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>