Rewrite Interface using Singleton Pattern
Change the Interface to require a Singleton.  This enables slightly
cleaner code as it now means that conformance is defined against an
Interface as opposed to against a Bundle.  This now requires an explicit
type member as opposed to a type parameter.

This entirely destroys the ParametricInterfaceSpec.  I need to just
rewrite this given the new approach.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>