Add InterfaceGenerator, update Parametric Example
Add a new concept of an "InterfaceGenerator".  This behaves like an
Interface, but is _not_ a Singleton.  This can then be used in a
pattern, shown in ParametricInterfaceSpec, to generate (statically)
different variations of an Interface.

This is all currently very manual.  A user needs to explicitly define
all possible Interfaces that they want from a given InterfaceGenerator.
This is also missing features related to how to lookup an Interface
given a Key in the Package.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>