Fix module instantiation for clock/reset
Use a dummy clock and reset when instantiating a conforming module.
This is necessary if the conforming module is a Module (which has the
implicit clock and reset ports) as opposed to a RawModule (which does
not).  This is fine because all ports of the conforming module are
hooked up using a DataView which will error if the clock/rest ports are
not overridden.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>